refact(Auth): reorganize redux files
