+++ src/components/layouts/main-menu/logout-item.tsx
@@ -6,8 +6,7 @@ import { Redirect } from 'react-router-dom';
 import { MenuItem } from 'react-bootstrap';
 
 import { addFlashMessage } from 'actions/flash-messages.js';
-import { logoutUser } from 'actions/auth.js';
-import { selectIsAuthorized } from 'selectors/auth.js';
+import { logoutUser, selectIsAuthorized } from 'services/auth';
 
 interface StateProps {
   isAuthorized?: boolean;
