+++ src/components/login/login.jsx
@@ -5,7 +5,7 @@ import { Redirect } from 'react-router-dom';
 import { Panel } from 'react-bootstrap'
 
 import { addFlashMessage } from 'actions/flash-messages.js';
-import { loginUser } from 'actions/auth.js';
+import { loginUser } from 'services/auth';
 import { prepareSubmissionError } from 'utils/errors';
 
 import Form from './form.jsx'
