+++ src/reducers/current-organization.js
@@ -1,7 +1,7 @@
 import { handleActions, combineActions } from 'redux-actions';
 
 import { setCurrentOrganization, clearCurrentOrganization } from 'actions/organizations.js';
-import { restoreSession } from 'actions/auth.js';
+import { restoreSession } from 'services/auth';
 
 const defaultState = {
   id:   null,
